CiteFuse

A function to runSNF for CITE seq data

CiteFuse(
  sce,
  altExp_name = "ADT",
  W_list = NULL,
  gene_select = TRUE,
  dist_cal_RNA = "correlation",
  dist_cal_ADT = "propr",
  ADT_subset = NULL,
  K_knn = 20,
  t = 20,
  metadata_names = NULL,
  verbose = TRUE
)

Arguments

sce
a SingleCellExperiment

altExp_name
expression name of ADT matrix

W_list
affinity list, if it is NULL, the function will calculate it.

gene_select
whether highly variable genes will be selected for RNA-seq to calcualte simlarity matrix

dist_cal_RNA
similarity metrics used for RNA matrix

dist_cal_ADT
similarity metrics used for ADT matrix

ADT_subset
A vector indicates the subset that will be used.

K_knn
Number of nearest neighbours

t
Number of iterations for the diffusion process.

metadata_names
A vector indicates the names of metadata returned

verbose
whether print out the process

Value

A SingleCellExperiment object with fused matrix results stored

Examples

data("sce_ctcl_subset", package = "CiteFuse")
sce_ctcl_subset <- CiteFuse(sce_ctcl_subset)

#> Calculating affinity matrix 
#> Performing SNF