readFrom10X

A function to read the data from 10X

readFrom10X(
  dir,
  type = c("auto", "sparse", "HDF5"),
  feature_named_by = c("gene_id", "gene_symbol"),
  filter_features = TRUE
)

Arguments

dir
A character indicates the directory of the 10X files

type
A character indicates the format of the data, sparse or HDF5

feature_named_by
A character indicates whehter the genes will be named by gene_id or gene_symbol

filter_features
A logical input indicates whether the features with all zeros will be removed

Value

a SingleCellExperiment object

Examples

tmpdir <- tempdir()
tenXdata <- "http://cf.10xgenomics.com/samples/cell-exp/3.1.0/connect_5k_pbmc_NGSC3_ch1/"
file <- "connect_5k_pbmc_NGSC3_ch1_filtered_feature_bc_matrix.tar.gz"
download.file(paste0(tenXdata, file),file.path(tmpdir, file))
untar(file.path(tmpdir,file),
      exdir = tmpdir)
sce_citeseq_10X <- readFrom10X(file.path(tmpdir,
"filtered_feature_bc_matrix/"))
sce_citeseq_10X

#> class: SingleCellExperiment 
#> dim: 33538 3996 
#> metadata(0):
#> assays(1): counts
#> rownames(33538): ENSG00000243485 ENSG00000237613 ... ENSG00000277475
#>   ENSG00000268674
#> rowData names(3): gene_id gene_symbol gene_type
#> colnames(3996): AAACCCAAGAGGTCAC-1 AAACGAAAGATGGTCG-1 ...
#>   TTTGTTGGTTCAGCTA-1 TTTGTTGTCCTTCAGC-1
#> colData names(0):
#> reducedDimNames(0):
#> altExpNames(1): ADT